+++ src/main/java/insane96mcp/enhancedai/modules/mobs/targeting/EASpiderTargetGoal.java
@@ -6,8 +6,9 @@ import net.minecraft.world.entity.ai.targeting.TargetingConditions;
 import net.minecraft.world.entity.monster.Spider;
 
 public class EASpiderTargetGoal<T extends LivingEntity> extends EANearestAttackableTarget<T> {
-    public EASpiderTargetGoal(Spider goalOwnerIn, Class<T> targetClassIn, boolean checkSight, boolean nearbyOnlyIn, TargetingConditions targetingConditions) {
+    public EASpiderTargetGoal(Spider goalOwnerIn, Class<T> targetClassIn, boolean checkSight, boolean nearbyOnlyIn, TargetingConditions targetingConditions, int unseenMemoryTicks) {
         super(goalOwnerIn, targetClassIn, checkSight, nearbyOnlyIn, targetingConditions);
+        this.unseenMemoryTicks = unseenMemoryTicks;
     }
 
     /**